+++ src/plugin/fixes/apply-suggestions.test.ts
