+++ src/plugin/fixes/disable-per-line.ts
@@ -1,9 +1,9 @@
 import { Linter, Rule } from 'eslint';
 import type { Comment } from 'estree';
-import { TransformContext } from '../../plugin/index.js';
 import { groupBy, unique } from '../../util/array.js';
 import { DisableComment, mergeRuleIdsAndDescription, parseDisableComment, toCommentText } from '../../util/eslint.js';
 import { notEmpty } from '../../util/type-check.js';
+import { TransformContext } from '../index.js';
 
 export type TransformToDisablePerLineArgs = {
   description?: string;
