+++ src/plugin/fixes/make-fixable-and-fix.ts
@@ -1,8 +1,8 @@
 import { Linter, Rule, SourceCode } from 'eslint';
 import { traverse } from 'estraverse';
 import type { Node } from 'estree';
-import { TransformContext } from '../../plugin/index.js';
 import { unreachable } from '../../util/type-check.js';
+import { TransformContext } from '../index.js';
 
 export type FixableMaker = (
   message: Linter.LintMessage,
