+++ src/plugin/index.ts
@@ -1,6 +1,4 @@
 import { Linter, Rule, SourceCode } from 'eslint';
-import { preferAdditionShorthandRule } from './prefer-addition-shorthand-rule.js';
-import { transformRule, type TransformRuleOption } from './transform-rule.js';
 import {
   type FixableMaker,
   type SuggestionFilter,
@@ -8,7 +6,9 @@ import {
   type TransformToDisablePerFileArgs,
   type TransformToDisablePerLineArgs,
   type TransformToMakeFixableAndFixArgs,
-} from './transforms/index.js';
+} from './fixes/index.js';
+import { preferAdditionShorthandRule } from './prefer-addition-shorthand-rule.js';
+import { transformRule, type TransformRuleOption } from './transform-rule.js';
 
 export { TransformRuleOption, type FixableMaker, type SuggestionFilter };
